Adds support for hex, ex_doc, and Elixir
Elixir is a popular alternative on the BEAM, and by supporting those
users as part of the RAIK ecosystem, we can bring Riak to even more
projects.

To generate documents, run `rebar ex_doc`. This command builds
documentation and places it in a `doc` folder. Open `doc/index.html` to
view the generated documentation.

I have set up the framework so that we can publish this to hex and get
more exposure. The documentation should be pointed to hex instead of the
GitHub repo if we publish there.